requirements.txt
Django>=4.2.7
graphene-django>=3.1.5
django-filter>=23.5
